fixed sign up and login pages
