+++ poke-app/app/SignUp.tsx
@@ -4,7 +4,8 @@ import { useNavigation } from '@react-navigation/native';
 import ParallaxScrollView from '@/components/ParallaxScrollView';
 import { ThemedText } from '@/components/ThemedText';
 import { ThemedView } from '@/components/ThemedView';
-export default function HomeScreen() {
+
+export default function SignupScreen() {
     return (
         <ThemedView style={styles.titleContainer}>
           <ThemedText type="title">Welcome!</ThemedText>
