+++ poke-app/app/_layout.tsx
@@ -28,8 +28,9 @@ export default function RootLayout() {
 
   return (
     <ThemeProvider value={colorScheme === 'dark' ? DarkTheme : DefaultTheme}>
-      <Stack>
-        <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
+      <Stack initialRouteName='index' >
+        <Stack.Screen name="index" options={{ headerShown: false }}/>
+        <Stack.Screen name="(tabs)"  options={{ headerShown: false }}/>
         <Stack.Screen name="+not-found" />
       </Stack>
     </ThemeProvider>
